Modify module to use persistent storage (rather than module globals), and create concept of storage solely for configuration purposes (cfg->configPool).
This new memory region allows us to allocate whatever we need during server configuration.  At the end of server configuration, this temporary configuration block is removed.

NOTE: THIS IS VERY DIFFERENT THAN APR POOL 'pconf'.  APR pool 'pconf' (as passed to the post_config_handler) is configuration-specific data that is persisted for the lifetime of the server.  But anything stored in memory pool cfg->configPool is deleted when all configuration handlers have been called, allowing for use of memory only during the configuration process.