Unit test framework for Apache Providers.
Note: This is just a placeholder at this time.  If you look at server_test.cpp, it doesn't actually do anything.  But the unit test does build/run successfully, and this gives me a framework that I can build upon.

NOTE: I'm having problems with 'tf unshelve' throwing an exception, causing the unshelve to fail. For testing, I needed to copy over from Windows (YUCK)! We may need to look into this if it continues to be a problem ...