Recognize if Apache restarts and remap the memory mapped file. This actually occurs by mapping the memory map for each individual operation, then closing the memory map afterwards. In this way, if Apache is restarted, we automatically get the new memory map.
Since the provider tends to store temporary data in the memory map, a new memory map forces temporary data to be recomputed. In this way, the provider "just works" in case Apache is restarted.

I also fixed a minor bug in the Apache_HTTPDVirtualHostCertficate enumeration: If we had a problem computing expiration information for a certificate, we used to abort the enumeration. Now we still enumerate but don't supply expiration information. In this case, we'll retry determining expiration information the next time we get the opportunity.