Apache configuration script to configure and unconfigure Apache Server; updates to installer for latest requirements from PM
Specifically:

1. On installation, do not configure or restart apache (unless --restart-deps is specified on command line)

2. On removal, don't touch Apache configuration if agent is already not configured (and don't restart server either). If agent is configured, then remove Apache configuration and restart Apache server.

3. Default bundle behavior should not be "extract" anymore; it's now a specific option.

This fix provides a new script in the Apache provider bin directory:

/opt/microsoft/apache-cimprov/bin/apache_config.sh

Run this with -c to configure Apache after installation, or run with -u to unconfigure Apache prior to removal of the agent.

IMPORTANT NOTE TO TEST:

Due to these changes, if you simply install the provider and try to enumerate, enumerations will fail unless the bundle was installed with --restart-deps or you run 'apache_config.sh -c' first.